# Conflicts:
#	scientific_library/tvb/simulator/integrators.py
#	scientific_library/tvb/simulator/simulator.py
#	scientific_library/tvb/tests/library/simulator/simulator_test.py

+++ framework_tvb/tvb/core/services/operation_service.py
@@ -43,11 +43,13 @@ import sys
 import uuid
 import zipfile
 from copy import copy
+from inspect import isclass
+
 from tvb.basic.exceptions import TVBException
 from tvb.basic.logger.builder import get_logger
 from tvb.basic.neotraits.api import Range
 from tvb.basic.profile import TvbProfile
-from tvb.config import choices, MEASURE_METRICS_MODULE, MEASURE_METRICS_CLASS, MEASURE_METRICS_MODEL_CLASS
+from tvb.config import MEASURE_METRICS_MODULE, MEASURE_METRICS_CLASS, MEASURE_METRICS_MODEL_CLASS, ALGORITHMS
 from tvb.core.adapters import constants
 from tvb.core.adapters.abcadapter import ABCAdapter, AdapterLaunchModeEnum
 from tvb.core.adapters.exceptions import LaunchException
@@ -131,8 +133,6 @@ class OperationService:
             generic_metadata.operation_tag = submit_data[DataTypeMetaData.KEY_OPERATION_TAG]
         if DataTypeMetaData.KEY_TAG_1 in submit_data:
             generic_metadata.user_tag_1 = submit_data[DataTypeMetaData.KEY_TAG_1]
-        if operation_group is not None:
-            generic_metadata.user_tag_3 = operation_group.name
         generic_metadata.fill_from(current_ga)
         return generic_metadata
 
@@ -170,7 +170,7 @@ class OperationService:
 
         view_model = get_class_by_name("{}.{}".format(MEASURE_METRICS_MODULE, MEASURE_METRICS_MODEL_CLASS))()
         view_model.time_series = time_series_index.gid
-        view_model.algorithms = tuple(choices.values())
+        view_model.algorithms = tuple(ALGORITHMS.keys())
         view_model.generic_attributes = ga
 
         parent_burst = dao.get_generic_entity(BurstConfiguration, time_series_index.fk_parent_burst, 'gid')[0]
@@ -289,6 +289,20 @@ class OperationService:
             available_space = disk_space_per_user - pending_op_disk_space - user_disk_space
 
             view_model = adapter_instance.load_view_model(operation)
+            try:
+                form = adapter_instance.get_form()
+                form = form() if isclass(form) else form
+                fields = form.get_upload_field_names()
+                project = dao.get_project_by_id(operation.fk_launched_in)
+                tmp_folder = self.file_helper.get_project_folder(project, self.file_helper.TEMP_FOLDER)
+                for upload_field in fields:
+                    if hasattr(view_model, upload_field):
+                        file = getattr(view_model, upload_field)
+                        if file.startswith(tmp_folder):
+                            temp_files.append(file)
+            except AttributeError:
+                # Skip if we don't have upload fields on current form
+                pass
             result_msg, nr_datatypes = adapter_instance._prelaunch(operation, view_model, unique_id, available_space)
             operation = dao.get_operation_by_id(operation.id)
             # Update DB stored kwargs for search purposes, to contain only valuable params (no unselected options)
@@ -296,13 +310,6 @@ class OperationService:
             dao.store_entity(operation)
 
             self._update_vm_generic_operation_tag(view_model, operation)
-
-            adapter_form = adapter_instance.get_form()
-            try:
-                temp_files = adapter_form.temporary_files
-            except AttributeError:
-                pass
-
             self._remove_files(temp_files)
 
         except zipfile.BadZipfile as excep:
@@ -380,15 +387,16 @@ class OperationService:
         Currently used to delete temporary files created during an operation.
         """
         for pth in file_list:
-            pth = str(pth)
-            try:
-                if os.path.exists(pth) and os.path.isfile(pth):
-                    os.remove(pth)
-                    self.logger.debug("We no longer need file:" + pth + " => deleted")
-                else:
-                    self.logger.warning("Trying to remove not existent file:" + pth)
-            except OSError:
-                self.logger.exception("Could not cleanup file!")
+            if pth is not None:
+                pth = str(pth)
+                try:
+                    if os.path.exists(pth) and os.path.isfile(pth):
+                        os.remove(pth)
+                        self.logger.debug("We no longer need file:" + pth + " => deleted")
+                    else:
+                        self.logger.warning("Trying to remove not existent file:" + pth)
+                except OSError:
+                    self.logger.exception("Could not cleanup file!")
 
     @staticmethod
     def _range_name(range_no):
